refactor: Add separate bundler for performance views
