+++ hrms/hr/doctype/appraisal/appraisal.js
@@ -1,6 +1,5 @@
 // Copyright (c) 2015, Frappe Technologies Pvt. Ltd. and Contributors
 // License: GNU General Public License v3. See license.txt
-frappe.provide("hrms");
 
 frappe.ui.form.on("Appraisal", {
 	refresh: function(frm) {
@@ -16,6 +15,8 @@ frappe.ui.form.on("Appraisal", {
 			frm.trigger("show_feedbacks");
 			frm.trigger("setup_chart");
 		}
+
+		frm.sidebar.image_wrapper.find(".sidebar-image-actions").addClass("hide");
 	},
 
 	kra_template: function(frm) {
@@ -44,11 +45,13 @@ frappe.ui.form.on("Appraisal", {
 	},
 
 	show_feedbacks: function(frm) {
-		const feedbacks = new hrms.Feedback({
-			frm: frm,
-			feedback_wrapper: $(frm.fields_dict.feedback_html.wrapper),
+		frappe.require("performance.bundle.js", () => {
+			const feedbacks = new hrms.EmployeeFeedback({
+				frm: frm,
+				feedback_wrapper: $(frm.fields_dict.feedback_html.wrapper),
+			});
+			feedbacks.refresh();
 		});
-		feedbacks.refresh();
 	},
 
 	setup_chart: function(frm) {
@@ -79,210 +82,11 @@ frappe.ui.form.on("Appraisal", {
 						}
 					]
 				},
-
 				title: __("Scores"),
 				height: 300,
-				colors: ["#4CA746", "#98D85B"],
-				type: "bar"
+				type: "bar",
+				colors: ["blue", "green"]
 			});
 		}
 	}
-
 });
-
-hrms.Feedback = class Feedback {
-	constructor(opts) {
-		$.extend(this, opts);
-	}
-
-	refresh() {
-		var me = this;
-		this.feedback_wrapper.find(".notes-section").remove();
-
-		frappe.call({
-			method: "hrms.hr.doctype.appraisal.appraisal.get_feedbacks",
-			args: {
-				employee: me.frm.doc.employee
-			},
-			callback: function(r) {
-				if (!r.exc) {
-					const feedbacks = r.message || [];
-					feedbacks.sort(
-						function(a, b) {
-							return new Date(b.added_on) - new Date(a.added_on);
-						}
-					);
-
-					const feedback_html = frappe.render_template("feedback", {
-						feedbacks: feedbacks
-					});
-					$(feedback_html).appendTo(me.feedback_wrapper);
-
-					$(".notes-section").find(".new-note-btn").on("click", function() {
-						me.add_feedback();
-					});
-
-					$(".notes-section").find(".edit-note-btn").on("click", function() {
-						me.edit_feedback();
-					});
-
-					$(".notes-section").find(".delete-note-btn").on("click", function() {
-						me.delete_feedback();
-					});
-				}
-			}
-		});
-	}
-
-
-	add_feedback() {
-		const me = this;
-		const _add_feedback = () => {
-			this.data = [];
-			const dialog = new frappe.ui.Dialog({
-				title: __("Add Feedback"),
-				fields: [
-					{
-						"label": "Feedback",
-						"fieldname": "feedback",
-						"fieldtype": "Text Editor",
-						"reqd": 1,
-						"enable_mentions": true,
-					},
-					{
-						"label": "KRA Rating",
-						"fieldtype": "Table",
-						"fieldname": "kra_rating",
-						"cannot_add_rows": true,
-						"data": this.data,
-						get_data: () => {
-							return this.data;
-						},
-						"fields": [
-							{
-								"fieldname": "kra",
-								"fieldtype": "Link",
-								"in_list_view": 1,
-								"label": "KRA",
-								"options": "KRA",
-								"reqd": 1
-							},
-							{
-								"fieldname": "per_weightage",
-								"fieldtype": "Percent",
-								"in_list_view": 1,
-								"label": "Weightage"
-							},
-							{
-								"fieldname": "rating",
-								"fieldtype": "Rating",
-								"in_list_view": 1,
-								"label": "Rating"
-							}
-						]
-					}
-				],
-				primary_action: function() {
-					var data = dialog.get_values();
-					frappe.call({
-						method: "add_feedback",
-						doc: me.frm.doc,
-						args: {
-							feedback: data.feedback,
-							kra_rating: data.kra_rating
-						},
-						freeze: true,
-						callback: function(r) {
-							if (!r.exc) {
-								me.frm.refresh_field("feedbacks_table");
-								me.refresh();
-							}
-							dialog.hide();
-						}
-					});
-				},
-				primary_action_label: __("Add")
-			});
-
-			frappe.call({
-				"method": "hrms.hr.doctype.performance_feedback.performance_feedback.get_kra",
-				args: {
-					employee: me.frm.doc.employee
-				},
-				callback: function(data) {
-					data.message.goals.forEach((item) => {
-						dialog.fields_dict.kra_rating.df.data.push({
-							"kra": item.kra,
-							"per_weightage": item.per_weightage
-						});
-						this.data = dialog.fields_dict.kra_rating.df.data;
-						dialog.fields_dict.kra_rating.grid.refresh();
-					})
-				}
-			});
-			dialog.show();
-		};
-		$(".new-note-btn").click(_add_feedback);
-	}
-
-	edit_feedback(edit_btn) {
-		const me = this;
-		const row = $(edit_btn).closest(".comment-content");
-		const row_id = row.attr("name");
-		const row_content = $(row).find(".content").html();
-		if (row_content) {
-			const d = new frappe.ui.Dialog({
-				title: __("Edit Feedback"),
-				fields: [
-					{
-						"label": "Feedback",
-						"fieldname": "feedback",
-						"fieldtype": "Text Editor",
-						"reqd": 1,
-						"enable_mentions": true,
-					}
-				],
-				primary_action: function() {
-					var data = d.get_values();
-					frappe.call({
-						method: "edit_feedback",
-						doc: me.frm.doc,
-						args: {
-							feedback: data.feedback,
-							row_id: row_id
-						},
-						freeze: true,
-						callback: function(r) {
-							if (!r.exc) {
-								me.frm.refresh_field("feedbacks_table");
-								me.refresh();
-							}
-							d.hide();
-						}
-					});
-				},
-				primary_action_label: __("Done")
-			});
-			d.show();
-		}
-	}
-
-	delete_feedback(delete_btn) {
-		const me = this;
-		const row_id = $(delete_btn).closest(".comment-content").attr("name");
-		frappe.call({
-			method: "delete_feedback",
-			doc: me.frm.doc,
-			args: {
-				row_id: row_id
-			},
-			freeze: true,
-			callback: function(r) {
-				if (!r.exc) {
-					me.frm.refresh_field("feedbacks_table");
-					me.refresh();
-				}
-			}
-		});
-	}
-};
