+++ hrms/public/js/templates/rating.html
@@ -1,0 +1,35 @@
+<div class="d-flex flex-col">
+	<div class="rating {{ for_summary ? 'ratings-pill' : ''}}">
+		{% for (let i = 1; i <= number_of_stars; i++) { %}
+			{% if (i <= average_rating) { %}
+				{% right_class = 'star-click'; %}
+			{% } else { %}
+				{% right_class = ''; %}
+			{% } %}
+
+			{% if ((i <= average_rating) || ((i - 0.5) == average_rating)) { %}
+				{% left_class = 'star-click'; %}
+			{% } else { %}
+				{% left_class = ''; %}
+			{% } %}
+
+			<svg class="icon icon-md" data-rating={{i}} viewBox="0 0 24 24" fill="none">
+				<path class="right-half {{ right_class }}" d="M11.9987 3.00011C12.177 3.00011 12.3554 3.09303 12.4471 3.27888L14.8213 8.09112C14.8941 8.23872 15.0349 8.34102 15.1978 8.3647L20.5069 9.13641C20.917 9.19602 21.0807 9.69992 20.7841 9.9892L16.9421 13.7354C16.8243 13.8503 16.7706 14.0157 16.7984 14.1779L17.7053 19.4674C17.7753 19.8759 17.3466 20.1874 16.9798 19.9945L12.2314 17.4973C12.1586 17.459 12.0786 17.4398 11.9987 17.4398V3.00011Z" fill="var(--star-fill)" stroke="var(--star-fill)"/>
+				<path class="left-half {{ left_class }}" d="M11.9987 3.00011C11.8207 3.00011 11.6428 3.09261 11.5509 3.27762L9.15562 8.09836C9.08253 8.24546 8.94185 8.34728 8.77927 8.37075L3.42887 9.14298C3.01771 9.20233 2.85405 9.70811 3.1525 9.99707L7.01978 13.7414C7.13858 13.8564 7.19283 14.0228 7.16469 14.1857L6.25116 19.4762C6.18071 19.8842 6.6083 20.1961 6.97531 20.0045L11.7672 17.5022C11.8397 17.4643 11.9192 17.4454 11.9987 17.4454V3.00011Z" fill="var(--star-fill)" stroke="var(--star-fill)"/>
+			</svg>
+		{% } %}
+	</div>
+	{% if (!for_summary) { %}
+		<p class="ml-3" style="line-height: 2;">
+			({{ average_rating }})
+		</p>
+	{% } %}
+</div>
+
+<style>
+	.ratings-pill {
+		background-color: var(--gray-100);
+		padding: .5rem 1rem;
+		border-radius: 66px;
+	}
+</style>
